+++ packages/vsce/src/utils/resourceUtils.ts
@@ -10,14 +10,22 @@
  */
 
 import { IResource } from "@zowe/cics-for-zowe-explorer-api";
-import { CicsCmciRestClient, getCache, getResource, ICMCIApiResponse, ICMCIResponseResultSummary, IResourceQueryParams, Utils } from "@zowe/cics-for-zowe-sdk";
+import {
+  getCache,
+  getResource,
+  ICMCIApiResponse,
+  ICMCIResponseResultSummary,
+  IResourceQueryParams,
+  putResource,
+} from "@zowe/cics-for-zowe-sdk";
 import { AuthOrder, IProfileLoaded } from "@zowe/imperative";
-import { extensions } from "vscode";
 import constants from "../constants/CICS.defaults";
-import { SessionHandler } from "../resources";
-import { CICSResourceContainerNode } from "../trees";
-import { CICSLogger } from "./CICSLogger";
 import { getErrorCode } from "./errorUtils";
+import { CICSExtensionError } from "../errors/CICSExtensionError";
+import { extensions } from "vscode";
+import { CICSLogger } from "./CICSLogger";
+import { CICSResourceContainerNode } from "../trees";
+import { SessionHandler } from "../resources";
 
 interface IReqParams {
   criteria?: string;
@@ -69,14 +77,14 @@ export async function runGetResource({ profileName, resourceName, regionName, ci
   } catch (error) {
     // Make sure the error is not caused by the ltpa token expiring
     if (getErrorCode(error) != constants.HTTP_ERROR_UNAUTHORIZED || !session.ISession?.tokenValue) {
-      throw error;
+      throw new CICSExtensionError({baseError: error});
     }
   }
 
   CICSLogger.debug("Retrying as validation of the LTPA token failed because the token has expired.");
   const newSession = buildNewSession(profile);
 
-  return await getResource(newSession, buildResourceParms(resourceName, regionName, cicsPlex, params), buildRequestOptions(), [
+  return getResource(newSession, buildResourceParms(resourceName, regionName, cicsPlex, params), buildRequestOptions(), [
     buildUserAgentHeader(),
   ]);
 }
@@ -107,8 +115,8 @@ export async function runGetCache(
     ]);
   } catch (error) {
     // Make sure the error is not caused by the ltpa token expiring
-    if (getErrorCode(error) !== constants.HTTP_ERROR_UNAUTHORIZED || !session.ISession?.tokenValue) {
-      throw error;
+    if (getErrorCode(error) !== constants.HTTP_ERROR_UNAUTHORIZED || !session.ISession.tokenValue) {
+      throw new CICSExtensionError({baseError: error});
     }
   }
 
@@ -116,7 +124,7 @@ export async function runGetCache(
 
   const newSession = buildNewSession(profile);
 
-  return await getCache(
+  return getCache(
     newSession,
     { cacheToken, startIndex, count, nodiscard: true, summonly: false },
     { failOnNoData: false, useCICSCmciRestError: true },
@@ -125,11 +133,6 @@ export async function runGetCache(
 }
 
 export async function runPutResource({ profileName, resourceName, regionName, cicsPlex, params }: IRunGetPutResourceParams, requestBody: any) {
-  const cmciResource = Utils.getResourceUri(resourceName, {
-    cicsPlex: cicsPlex,
-    regionName: regionName,
-    ...params,
-  });
 
   CICSLogger.debug(
     buildRequestLoggerString("PUT", resourceName, {
@@ -144,25 +147,27 @@ export async function runPutResource({ profileName, resourceName, regionName, ci
   const profile = SessionHandler.getInstance().getProfile(profileName);
   const session = SessionHandler.getInstance().getSession(profile);
 
+
   try {
     // First attempt
     if (!session.ISession?.tokenValue) {
       AuthOrder.makingRequestForToken(session.ISession);
     }
-
-    return await CicsCmciRestClient.putExpectParsedXml(session, cmciResource, [buildUserAgentHeader()], requestBody);
+    return await putResource(session, buildResourceParms(resourceName, regionName, cicsPlex, params)
+      , [buildUserAgentHeader()], requestBody, buildRequestOptions());
+    
   } catch (error) {
     // Make sure the error is not caused by the ltpa token expiring
-    if (getErrorCode(error) !== constants.HTTP_ERROR_UNAUTHORIZED || !session.ISession?.tokenValue) {
-      throw error;
+    if (getErrorCode(error) !== constants.HTTP_ERROR_UNAUTHORIZED || !session.ISession.tokenValue) {
+      throw new CICSExtensionError({ baseError: error });
     }
   }
 
   CICSLogger.debug("Retrying as validation of the LTPA token failed because the token has expired.");
-
   const newSession = buildNewSession(profile);
 
-  return await CicsCmciRestClient.putExpectParsedXml(newSession, cmciResource, [buildUserAgentHeader()], requestBody);
+  return putResource(newSession, buildResourceParms(resourceName, regionName, cicsPlex, params)
+      , [buildUserAgentHeader()], requestBody, buildRequestOptions());
 }
 
 export const buildResourceParms = (resourceName: string, regionName: string, cicsplexName: string, params: IReqParams) => {
@@ -194,7 +199,7 @@ export const buildNewSession = (profile: IProfileLoaded) => {
 export const buildRequestLoggerString = (
   method: "GET" | "PUT" | "POST",
   resourceName: string,
-  opts: { [key: string]: string | boolean | number } = {}
+  opts: { [key: string]: string | boolean | number; } = {}
 ): string => {
   let output = `${method.toUpperCase()} - Resource [${resourceName}]`;
   for (const [k, v] of Object.entries(opts)) {
@@ -235,7 +240,7 @@ export async function pollForCompleteAction<T extends IResource>(
   criteriaMetCallback(response);
 }
 
-export function buildUserAgentHeader(): { "User-Agent": string } {
+export function buildUserAgentHeader(): { "User-Agent": string; } {
   const zeId = `zowe.vscode-extension-for-zowe`;
   const cicsExtId = `zowe.cics-extension-for-zowe`;