+++ packages/sdk/src/rest/CicsCmciRestError.ts
@@ -10,21 +10,28 @@
  */
 
 import { ImperativeError } from "@zowe/imperative";
-import { ICMCIResponseResultSummary } from "../doc";
+import { ICMCIApiResponse, ICMCIResponseResultSummary } from "../doc";
+import { CicsCmciConstants } from "../constants";
+import { ICMCIResponseErrors } from "../doc/ICMCIResponseErrors";
 
 export class CicsCmciRestError extends ImperativeError {
   resultSummary: ICMCIResponseResultSummary;
+  errors: ICMCIResponseErrors;
 
   RESPONSE_1: number;
   RESPONSE_2: number;
   RESPONSE_1_ALT: string;
   RESPONSE_2_ALT: string;
+  FEEDBACKRESP: number;
+  FEEDBACKRESP_2: number;
+  FEEDBACK_ACTION: string;
+  EIBFN_ALT: string;
+  FEEDBACKRESP_ALT: string;
 
-  constructor(msg: string, resultSummary: ICMCIResponseResultSummary) {
-    super({
-      msg,
-    });
-    this.resultSummary = resultSummary;
+  constructor(msg: string , response: ICMCIApiResponse) {
+    super({msg});
+    this.resultSummary = response.response.resultsummary;
+    this.errors = response.response.errors;
     this.parseResultSummary();
   }
 
@@ -33,5 +40,10 @@ export class CicsCmciRestError extends ImperativeError {
     this.RESPONSE_2 = parseInt(this.resultSummary.api_response2);
     this.RESPONSE_1_ALT = this.resultSummary.api_response1_alt;
     this.RESPONSE_2_ALT = this.resultSummary.api_response2_alt;
+    this.FEEDBACKRESP = parseInt(this.errors?.feedback?.resp || CicsCmciConstants.DEFAULT_RESP_CODE);
+    this.FEEDBACKRESP_2 = parseInt(this.errors?.feedback?.resp2 || CicsCmciConstants.DEFAULT_RESP_CODE);
+    this.FEEDBACK_ACTION = this.errors?.feedback?.action;
+    this.FEEDBACKRESP_ALT = this.errors?.feedback?.resp_alt;
+    this.EIBFN_ALT = this.errors?.feedback?.eibfn_alt
   }
 }