+++ packages/sdk/CHANGELOG.md
@@ -2,11 +2,14 @@
 
 All notable changes to the IBM® CICS® Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Added PUT API [#453](https://github.com/zowe/cics-for-zowe-client/issues/453)
+
 ## `6.15.0`
 
 - Added CICS Remote File resource constant. [#416](https://github.com/zowe/cics-for-zowe-client/issues/416)
 - Added TS Queue and Shared TS Queue constants. [#415](https://github.com/zowe/cics-for-zowe-client/issues/415)
-- Enhancement: Added PUT API [#453](https://github.com/zowe/cics-for-zowe-client/issues/453)
 
 ## `6.14.0`