+++ packages/sdk/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the IBM® CICS® Plug-in for Zowe CLI will be documented in this file.
 
-## Recent Changes
+## `6.16.0`
 
 - Enhancement: Added putResource API to update CICS resources via CMCI. [#453](https://github.com/zowe/cics-for-zowe-client/issues/453)